+++ app/Http/Controllers/CommentController.php
@@ -11,27 +11,36 @@ class CommentController extends Controller
 {
     //
 
-    public function index(){
+    public function index()
+    {
 
 
     }
 
-    public function __construct(){
+    public function __construct()
+    {
 
     }
 
-    public function create(Request $request){
+    public function create(Request $request)
+    {
 
     }
 
-    public function store(Request $request){
+    public function store(Request $request)
+    {
 
-        if(!Auth::check()){
+        if (!Auth::check()) {
             return redirect()->route('login')->with('error', 'Please log in');
         }
 
-        if($request->user()->cant('create', Comment::class))
-            abort(401, 'Unauthorized operation'); 
+        if ($request->user()->cant('create', Comment::class))
+            abort(401, 'Unauthorized operation, contanct the admin');
+
+        // If the user has been blocked
+        if ($request->user()->hasRole('blocked')) {
+            return redirect()->route('blocked');
+        }
 
         $request->validate([
             'text' => 'required|max:255',
@@ -42,7 +51,7 @@ class CommentController extends Controller
         $comment = new Comment();
         $comment->text = $request->input('text');
         $comment->user_id = $request->user()->id;
-        $comment->article_id = $request->input('article_id'); 
+        $comment->article_id = $request->input('article_id');
         $comment->updated_at = $currentDateTime;
 
         $comment->save();
@@ -50,14 +59,20 @@ class CommentController extends Controller
         return back()->with('success', 'Your comment has been published');
     }
 
-    public function destroy(Request $request, Comment $comment){
+    public function destroy(Request $request, Comment $comment)
+    {
 
-        if(!Auth::check()){
-            return redirect()->route('login')->with('error', 'Please log in');           
+        if (!Auth::check()) {
+            return redirect()->route('login')->with('error', 'Please log in');
         }
 
-        if($request->user()->cant('delete', $comment))
-            abort(401, 'Unauthorized operation');
+        if ($request->user()->cant('delete', $comment))
+            abort(401, 'Unauthorized operation, contact the admin');
+
+        // If the user has been blocked
+        if ($request->user()->hasRole('blocked')) {
+            return redirect()->route('blocked');
+        }
 
         $comment->delete();
